Cleanup hu-seg
