+++ hu-segmentation/segment_with_classifier.py
@@ -1,3 +1,5 @@
+# CUDA_VISIBLE_DEVICES=2 python segment_with_classifier.py -t 1 -d kitti-360 -s 5 -o /project_data/ramanan/achakrav/4D-PLS/results/validation/val_preds_TS1_kitti360_updated_vocab -sd ../results/predictions/Kitti360/hlps_agnostic_regressor_TS1_updated_building
+# System imports
 import argparse
 import glob
 import os
@@ -6,11 +8,13 @@ import yaml
 
 sys.path.append("segment_classifier/")
 
+# Third-party imports
 import numpy as np
 from sklearn.cluster import DBSCAN
 import torch
 from tqdm import tqdm
 
+# Relative imports
 from segment_classifier.model.pointnet2 import PointNet2Classification
 from tree_utils import flatten_indices, flatten_labels, flatten_scores
 from utils import *
@@ -165,6 +169,7 @@ def parse_args():
     parser.add_argument("-s", "--sequence", help="Sequence", type=int, default=8)
     parser.add_argument("-o", "--objsem_folder", help="Folder with object and semantic predictions", type=str, default="/project_data/ramanan/mganesin/4D-PLS/test/4DPLS_original_params_original_repo_nframes1_1e-3_softmax/val_probs")
     parser.add_argument("-sd", "--save_dir", help="Save directory", type=str, default='test/LOSP')
+    parser.add_argument("-n", "--network", help="Which network to use", choices=["regressor", "classifier"], default="regressor")
     parser.add_argument("--ckpt", help="Checkpoint to load for segment classifier", type=str, default=None)
     parser.add_argument("--use-sem-features", help="Whether to use semantic features in classifier", action="store_true")
     parser.add_argument("--use-sem-refinement", help="Whether to use semantic refinement head", action="store_true")
@@ -176,37 +181,40 @@ if __name__ == '__main__':
     args = parse_args()
 
     if args.task_set == 0:
-        # unk_labels = [1, 2]
+        known_labels = [1, 2]
         unk_labels = [7]
     elif args.task_set == 1:
-        # unk_labels = [1, 2, 3]
+        if args.dataset == "semantic-kitti":
+            known_labels = [1, 2, 3]
+        else:
+            known_labels = [1, 2, 3, 6]
         unk_labels = [10]
     elif args.task_set == 2:
-        # unk_labels = [1, 2, 3, 4, 5]
+        if args.dataset == "semantic-kitti":
+            known_labels = [1, 2, 3, 4, 5]
+        else:
+            known_labels = [1, 2, 3, 4, 5, 9, 14, 15]
         unk_labels = [16]
     elif args.task_set == -1:
-        unk_labels = range(1, 9)
+        known_labels = range(1, 9)
+        unk_labels = [1]
     else:
         raise ValueError('Unknown task set: {}'.format(args.task_set))
 
     if args.use_sem_refinement:
         in_channels = 256 # 0
-        args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_focal_sem_feats_sem_refinement_weighted/checkpoints/epoch_200.pth"
-        # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/sem_feats_only_focal_sem_refinement_weighted_project_input_1024/checkpoints/epoch_200.pth"
-        # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_focal_sem_feats_sem_reifnement_weighted_new_indices/checkpoints/epoch_200.pth"
-        # args.ckpt = "/project_data/ramanan/mganesin/4D-PLS/results/checkpoints/xyz_sem_mean_refine/checkpoints/epoch_200.pth"
+        # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_focal_sem_feats_sem_refinement_weighted/checkpoints/epoch_200.pth"
     elif args.use_sem_features:
         in_channels = 256
-        # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/sem_xyz/checkpoints/epoch_200.pth"
-        args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_focal_sem_feats/checkpoints/epoch_50.pth"
+        # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_focal_sem_feats/checkpoints/epoch_50.pth"
     else:
         in_channels = 0
-        if args.task_set == -1:
-            args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_regression_MSE_LOSS_200_double_mlp_full_dataset/checkpoints/epoch_200.pth"
-        elif args.task_set == 1:
-            args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_regressor_TS1/checkpoints/epoch_200.pth"
-        else:
-            args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_regressor_TS2/checkpoints/epoch_200.pth"
+        # if args.task_set == -1:
+            # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_regression_MSE_LOSS_200_double_mlp_full_dataset/checkpoints/epoch_200.pth"
+        # elif args.task_set == 1:
+            # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_{}_TS1/checkpoints/epoch_200.pth".format(args.network)
+        # else:
+            # args.ckpt = "/project_data/ramanan/achakrav/4D-PLS/results/checkpoints/xyz_mean_{}_TS2/checkpoints/epoch_200.pth".format(args.network)
 
     # instantiate the segment classifier
     cfg = Config()
@@ -347,10 +355,14 @@ if __name__ == '__main__':
         else:
             semantic_features = None
 
-        if len(unk_labels) == 1:
-            mask = np.where(labels == unk_labels[0])
-        else:
-            mask = np.where(np.logical_and(labels > 0, labels < np.max(unk_labels)))
+
+        # if len(unk_labels) == 1:
+        #     mask = np.where(labels == unk_labels[0])
+        # else:
+        #     mask = np.where(np.logical_and(labels > 0, labels < np.max(unk_labels)))
+        mask = np.logical_and(labels > 0, labels == unk_labels[0])
+        for known_label in known_labels:
+            mask = np.logical_or(mask, labels == known_label)
 
         pts_velo_cs_objects = pts_velo_cs[mask]
         objectness_objects = objectness[mask]  # todo: change objectness_objects into a local variable
